updated cache to be better and more efficient

+++ utils/tools/cache.py
@@ -3,43 +3,12 @@ import datetime
 import uuid
 from io import BytesIO
 from disnake.ext import tasks
-from tinydb import TinyDB, Query, middlewares
-from tinydb.storages import JSONStorage
+import orjson
 
 from utils.tools.helpers import *
 from utils.tools.logger import logger
 from utils.tools.settings import settings
 
-
-# a middleware we can use to delay flushing till its needed. ie: dont write on every query
-class DelayedStorageMiddleware(middlewares.Middleware):
-	def __init__(self, storage_cls):
-		super().__init__(storage_cls)
-		self.cache = None
-
-	def read(self):
-		if self.cache is None:
-			self.cache = self.storage.read()
-		return self.cache
-
-	def write(self, data):
-		self.cache = data
-
-	# actually write data to disk
-	def flush(self):
-		self.storage.write(self.cache)
-
-	def close(self):
-		self.flush()
-		self.storage.close()
-
-# EXAMPLE CACHE ITEM
-# {
-# 	"key": "<the cache key>",
-# 	"file": "<filename>",
-# 	"timestamp": "<last time this file was visited>"
-# }
-CacheItem = Query()
 
 # currently set to 1 week timeout for cache files
 CACHE_FILE_TIMEOUT_MS = 1000 * 60 * 60 * 24 * 7
@@ -49,18 +18,55 @@ def get_timestamp(date=None):
 		date = datetime.datetime.now()
 	return int(datetime.datetime.timestamp(date) * 1000)
 
+class CacheItem(dict):
+	filename: str
+	timestamp: int
+	def __init__(self, json_data={}): # only to be used internally
+		for key in json_data:
+			self[key] = json_data[key]
+		
+	@classmethod
+	def create(cls, filename):
+		item = CacheItem()
+		item["filename"] = filename
+		item.update_timestamp()
+		return item
+	
+	@property
+	def filename(self):
+		return self["filename"]
+
+	@property
+	def timestamp(self):
+		return self["timestamp"]
+
+	def update_timestamp(self):
+		self["timestamp"] = get_timestamp()
+
+
 class Cache:
+	cache_data: typing.Dict[str, CacheItem]
 	def __init__(self, loop):
 		self.loop = loop
+		self.lock = asyncio.Lock(loop=self.loop)
 		self.cache_dir = settings.resource("cache/")
 		if not os.path.exists(self.cache_dir):
 			os.makedirs(self.cache_dir)
-		self.db = TinyDB(self.cache_dir + "_cache_db.json", storage=DelayedStorageMiddleware(JSONStorage))
-		self.lock = asyncio.Lock(loop=self.loop)
+		self.cache_data = {}
+		self.cache_index_filename = self.cache_dir + "_cache_index.json"
+		if os.path.exists(self.cache_index_filename):
+			with open(self.cache_index_filename, "rb") as f:
+				json_dict = orjson.loads(f.read())
+				for uri, item in json_dict.items():
+					self.cache_data[uri] = CacheItem(item)
+	
+	def _save_to_disk(self):
+		with open(self.cache_index_filename, "wb+") as f:
+			f.write(orjson.dumps(self.cache_data))
 	
 	@property
 	def size(self):
-		return len(self.db)
+		return len(self.cache_data)
 	
 	# Cleans up any old files and flushes the cache to disk
 	@tasks.loop(hours=4)
@@ -68,24 +74,25 @@ class Cache:
 		threshold = get_timestamp() - CACHE_FILE_TIMEOUT_MS
 		async with self.lock:
 			timer = SimpleTimer()
-			expired_items = self.db.search(CacheItem.timestamp < threshold)
-			expired_item_ids = list(map(lambda item: item.doc_id, expired_items))
-			for item in expired_items:
-				filename = self.cache_dir + item["file"]
-				if os.path.isfile(filename):
-					os.remove(filename)
-			self.db.remove(doc_ids=expired_item_ids)
-			self.db.storage.flush()
-			logger.info(f"{len(expired_item_ids)} culled from cache in {timer.miliseconds}ms. New cache size {self.size}")
+			removed_count = 0
+			for uri, item in list(self.cache_data.items()):
+				if item.timestamp < threshold:
+					filename = self.cache_dir + item.filename
+					if os.path.isfile(filename):
+						os.remove(filename)
+					removed_count += 1
+					del self.cache_data[uri]
+			self._save_to_disk()
+			logger.info(f"{removed_count} culled from cache in {timer.miliseconds}ms. New cache size {self.size}")
 
 	# Returns the filename of the cached url if it exists, otherwise None
 	async def get_filename(self, uri):
 		async with self.lock:
-			item = self.db.get(CacheItem.key == uri)
+			item = self.cache_data.get(uri)
 			if item is None:
 				return None
-			self.db.update({"timestamp": get_timestamp()}, doc_ids=[item.doc_id])
-			filename = self.cache_dir + item["file"]
+			item.update_timestamp()
+			filename = self.cache_dir + item.filename
 		if not os.path.isfile(filename):
 			return None
 		return filename
@@ -111,20 +118,16 @@ class Cache:
 	#creates a new entry in the cache and returns the filename of the new entry
 	async def new(self, uri, extension=None):
 		async with self.lock:
-			item = self.db.get(CacheItem.key == uri)
+			item = self.cache_data.get(uri)
 			if item is not None:
-				filename = self.cache_dir + item["file"]
+				filename = self.cache_dir + item.filename
 				if os.path.isfile(filename):
 					return filename
 			filename = str(uuid.uuid4())
 			if extension:
 				filename = f"{filename}.{extension}"
-			self.db.upsert({
-				"key": uri,
-				"file": filename,
-				"timestamp": get_timestamp()
-			}, CacheItem.key == uri)
-			self.db.storage.flush()
+			self.cache_data[uri] = CacheItem.create(filename)
+			self._save_to_disk()
 		return self.cache_dir + filename
 
 
@@ -148,9 +151,9 @@ class Cache:
 
 	async def remove(self, uri):
 		async with self.lock:
-			item = self.db.get(CacheItem.key == uri)
+			item = self.cache_data.get(uri)
 			if item is not None:
-				filename = self.cache_dir + item["file"]
+				filename = self.cache_dir + item.filename
 				if os.path.isfile(filename):
 					os.remove(filename)
-				self.db.remove(doc_ids=[item.doc_id])
+				del self.cache_data[uri]
